trueLL method and argument trueLL removed from fitting
The function trueLL  converted to a method, currently defined for abundance vectors and an object of fitsad-class

Arguments 'trueLL' (and thus 'dec.places') removed from all fitting functions that had it. Man pages corrected accordingly. This was doen because fitting with trueLL=TRUE changed directly the slot min of the fitsad object, which can break the object structure. TBI: another way to calculate loglik and IC values with loglikelihood as calculated by trueLL.

Utility functions not to be exported were transferred to utils.R. Currently only one function: is.whole.number